+++ LotComWatcher/Models/Services/NetworkService.cs
@@ -1,10 +1,16 @@
 using System.Net;
 using System.Net.Sockets;
 using System.Text;
+using LotComWatcher.Models.Datatypes;
+using LotComWatcher.Models.Enums;
+using LotComWatcher.Models.Extensions;
 
 namespace LotComWatcher.Models.Services;
 
-public sealed class NetworkService
+/// <summary>
+/// Provides methods to communicate with Scanners over TCP network connections.
+/// </summary>
+public static class NetworkService
 {
     /// <summary>
     /// Sets the default communication port for sending messages to the Cognex Scanners.
@@ -47,14 +53,6 @@ public sealed class NetworkService
         }
         // connection was established without exceptions
         return true;
-    }
-
-    /// <summary>
-    /// Creates a NetworkService object that enables communication over TCP networks.
-    /// </summary>
-    public NetworkService()
-    {
-        
     }
 
     /// <summary>
@@ -125,7 +123,7 @@ public sealed class NetworkService
     /// <returns>'true' if Message was successfully sent to ScannerAddress.</returns>
     /// <exception cref="SystemException"></exception>
     /// <exception cref="ArgumentException"></exception>
-    public async Task<bool> SendDataValidationError(IPAddress ScannerAddress, string LCDText, int Duration)
+    public static async Task<bool> SendDataValidationError(IPAddress ScannerAddress, string LCDText, int Duration)
     {
         // send Data Validation Failure and Send Alert DMCCs to the Scanner
         try
@@ -157,7 +155,7 @@ public sealed class NetworkService
     /// <returns>'true' if Message was successfully sent to ScannerAddress.</returns>
     /// <exception cref="SystemException"></exception>
     /// <exception cref="ArgumentException"></exception>
-    public async Task<bool> SendMissingPreviousScanError(IPAddress ScannerAddress, int Duration, List<string> PreviousProcess)
+    public static async Task<bool> SendMissingPreviousScanError(IPAddress ScannerAddress, int Duration, IEnumerable<int> PreviousProcess)
     {
         // send Data Validation Failure and Send Alert DMCCs to the Scanner
         try
@@ -165,7 +163,7 @@ public sealed class NetworkService
             bool Sent = SendMessage(ScannerAddress, "||>OUTPUT.DATAVALID-FAIL\r\n").Result;
             if (Sent)
             {
-                await SendMessage(ScannerAddress, $"||>UI.SEND-ALERT {Duration} 2 \"This Label was not scanned by {PreviousProcess[0]}. Basket is not valid for use.\"\r\n");
+                await SendMessage(ScannerAddress, $"||>UI.SEND-ALERT {Duration} 2 \"This Label was not scanned by {PreviousProcess.First()}. Basket is not valid for use.\"\r\n");
             }
         }
         catch (ArgumentException)
@@ -188,7 +186,7 @@ public sealed class NetworkService
     /// <returns>'true' if Message was successfully sent to ScannerAddress.</returns>
     /// <exception cref="SystemException"></exception>
     /// <exception cref="ArgumentException"></exception>
-    public async Task<bool> SendDuplicateScanError(IPAddress ScannerAddress, int Duration)
+    public static async Task<bool> SendDuplicateScanError(IPAddress ScannerAddress, int Duration)
     {
         // send Data Validation Failure and Send Alert DMCCs to the Scanner
         try
@@ -197,6 +195,67 @@ public sealed class NetworkService
             if (Sent)
             {
                 await SendMessage(ScannerAddress, $"||>UI.SEND-ALERT {Duration} 2 \"Duplicate Label scanned.\"\r\n");
+            }
+        }
+        catch (ArgumentException)
+        {
+            throw new SystemException("Could not establish a connection to the Scanner.");
+        }
+        catch (SystemException)
+        {
+            throw new SystemException("Failed to request a connection.");
+        }
+        return true;
+    }
+
+    /// <summary>
+    /// Attempts to send an Invalid Part error code to ScannerAddress.
+    /// Displays a notice on the Scanner's LCD screen for Duration seconds.
+    /// </summary>
+    /// <param name="ScannerAddress"></param>
+    /// <param name="Duration"></param>
+    /// <returns>'true' if Message was successfully sent to ScannerAddress.</returns>
+    /// <exception cref="SystemException"></exception>
+    /// <exception cref="ArgumentException"></exception>
+    public static async Task<bool> SendInvalidPartError(IPAddress ScannerAddress, int Duration)
+    {
+        // send Data Validation Failure and Send Alert DMCCs to the Scanner
+        try
+        {
+            bool Sent = SendMessage(ScannerAddress, "||>OUTPUT.DATAVALID-FAIL\r\n").Result;
+            if (Sent)
+            {
+                await SendMessage(ScannerAddress, $"||>UI.SEND-ALERT {Duration} 2 \"Invalid Part Label scanned.\"\r\n");
+            }
+        }
+        catch (ArgumentException)
+        {
+            throw new SystemException("Could not establish a connection to the Scanner.");
+        }
+        catch (SystemException)
+        {
+            throw new SystemException("Failed to request a connection.");
+        }
+        return true;
+    }
+
+    /// <summary>
+    /// Sends a Scan Validation Failure message to the Scanner that produced New.
+    /// </summary>
+    /// <param name="New"></param>
+    /// <param name="Duration"></param>
+    /// <param name="Fault"></param>
+    /// <returns></returns>
+    /// <exception cref="SystemException"></exception>
+    public static async Task<bool> SendScanValidationError(ScanOutput New, int Duration, ValidationFailure Fault)
+    {
+        // send Data Validation Failure and Send Alert DMCCs to the Scanner
+        try
+        {
+            bool Sent = await SendMessage(New.ScanAddress, "||>OUTPUT.DATAVALID-FAIL\r\n");
+            if (Sent)
+            {
+                await SendMessage(New.ScanAddress, $"||>UI.SEND-ALERT {Duration} 2 \"{ValidationFailureExtensions.ToMessage(Fault)}\"\r\n");
             }
         }
         catch (ArgumentException)