+++ LotComWatcher/Models/Services/ScanValidationService.cs
@@ -1,7 +1,6 @@
-using LotCom.Database;
-using LotCom.Enums;
-using LotCom.Types;
+using LotCom.Core.Models;
 using LotComWatcher.Models.Datatypes;
+using LotComWatcher.Models.Enums;
 
 namespace LotComWatcher.Models.Services;
 
@@ -11,144 +10,153 @@ namespace LotComWatcher.Models.Services;
 public static class ScanValidationService
 {
     /// <summary>
-    /// Compares two Dates (one from new ScanOutput and one from existing Scan). 
-    /// Returns whether the ExistingDate is within the passed RangeInDays after NewDate.
-    /// </summary>
-    /// <param name="NewDate"></param>
-    /// <param name="ExistingDate"></param>
-    /// <param name="RangeInDays"></param>
-    /// <returns></returns>
-    private static bool CompareDatesAsRange(DateTime NewDate, DateTime ExistingDate, int RangeInDays)
-    {
-        TimeSpan ElapsedTime = NewDate.Subtract(ExistingDate);
-        if (ElapsedTime.Days < 0 || ElapsedTime.Days > RangeInDays)
-        {
-            return false;
-        }
-        else
-        {
-            return true;
-        }
-    }
-
-    /// <summary>
-    /// Compares serial data of a ScanOutput object and an existing Scan object.
-    /// </summary>
-    /// <param name="New"></param>
-    /// <param name="Existing"></param>
-    /// <returns></returns>
-    private static bool CompareAsSameProcess(ScanOutput New, Scan Existing)
-    {
-        // create SerialNumber objects from the New and Existing objects
-        SerialNumber NewSerial;
-        SerialNumber ExistingSerial;
-        try
-        {
-            NewSerial = New.GetSerialNumber();
-            ExistingSerial = Existing.GetSerialNumber();
-        }
-        catch (FormatException)
-        {
-            throw new FormatException("Could not create a SerialNumber for the New and/or Existing Scan.");
-        }
-        // compare the SerialNumbers and Dates
-        if
-        (
-            NewSerial.GetFormattedValue().Equals(ExistingSerial.GetFormattedValue()) &&
-            NewSerial.Part.PartNumber.Equals(ExistingSerial.Part.PartNumber) &&
-            New.ProductionDate.Equals(Existing.ProductionDate)
-        )
-        {
-            return true;
-        }
-        // comparison does not match
-        return false;
-    }
-
-    /// <summary>
-    /// Compares a new ScanOutput and an existing Scan for a Label match.
-    /// </summary>
-    /// <param name="New"></param>
-    /// <param name="Existing"></param>
-    /// <returns></returns>
-    private static bool CompareAsPreviousProcess(ScanOutput New, Scan Existing)
-    {
-        // create SerialNumber objects from the New and Existing objects
-        string NewSerial;
-        string ExistingSerial;
-        try
-        {
-            NewSerial = New.GetSerialNumber().GetFormattedValue();
-            ExistingSerial = Existing.GetSerialNumber().GetFormattedValue();
-        }
-        catch (FormatException)
-        {
-            throw new FormatException("Could not create a SerialNumber for the New and/or Existing Scan.");
-        }
-        // compare SerialNumbers for a match
-        bool SerialNumberOK;
-        if
-        (
-            New.Process.Type == ProcessType.Machining &&
-            New.Process.PreviousProcesses![0].Equals("4470-DC-Deburr")
-        )
-        {
-            // if New is a Scan of Label from a machining Process, the serial number has changed
-            // need to check if New's DeburrJBKNumber matches Existing's serial number
-            SerialNumberOK = New.VariableFields.DeburrJBKNumber!.Formatted.Equals(ExistingSerial);
-        }
-        else
-        {
-            // serials must match; change only occurs from Deburr -> Machining processes
-            SerialNumberOK = NewSerial.Equals(ExistingSerial);
-        }
-        // compare the New and Existing Model Numbers
-        bool ModelNumberOK = New.Part.ModelNumber.Code.Equals(Existing.Part.ModelNumber.Code);
-        // compare the New and Existing Dates within a 60 day range
-        bool DateRangeOK = CompareDatesAsRange(New.ProductionDate, Existing.ProductionDate, 60);
-        // confirm that each of the three conditions are OK
-        if (SerialNumberOK && ModelNumberOK && DateRangeOK)
-        {
-            return true;
-        }
-        // comparison does not match
-        return false;
-    }
-
-    /// <summary>
     /// Validates that a matching Scan exists in the previous Process' database table.
     /// </summary>
     /// <param name="New"></param>
     /// <returns></returns>
-    public static bool ValidatePreviousProcess(ScanOutput New, IEnumerable<Scan> DatabaseSet)
+    private static async Task<bool> ValidatePreviousProcessScanExists(ScanOutput New, IEnumerable<Scan> DbSet)
     {
-        // compare the New data to each of the DatabaseSet entries
-        foreach (Scan _scan in DatabaseSet)
+        // confirm that the Label Process needs to have a Previous Process performed
+        if (!New.LabelProcess.HasPreviousProcess())
         {
-            if (CompareAsPreviousProcess(New, _scan))
-            {
-                return true;
-            }
+            return true;
         }
-        return false;
-    }
-
-    /// <summary>
-    /// Checks if New is unique within the passed set of Database entries. 
-    /// </summary>
-    /// <param name="New"></param>
-    /// <param name="DatabaseSet"></param>
-    /// <returns></returns>
-    public static bool ValidateUniqueScan(ScanOutput New, IEnumerable<Scan> DatabaseSet)
-    {
-        // compare New data to each of the DatabaseSet entries
-        foreach (Scan _scan in DatabaseSet)
+        return await Task.Run(() =>
         {
-            if (CompareAsSameProcess(New, _scan))
+            // compare the New data to each of the DatabaseSet entries
+            Scan? Previous = DbSet
+                .Where
+                (x => New
+                    .ToScan()
+                    .IsFromPreviousProcess(x)
+                )
+                .FirstOrDefault();
+            if (Previous is null)
             {
                 return false;
             }
+            else
+            {
+                return true;
+            }
+        });
+    }
+
+    /// <summary>
+    /// Checks if New is unique within the context of its Process. 
+    /// </summary>
+    /// <param name="New"></param>
+    /// <returns>true if the Scan is Unique; false if not.</returns>
+    private static async Task<bool> ValidateAsUniqueScan(ScanOutput New, IEnumerable<Scan> DbSet)
+    {
+        return await Task.Run(() =>
+        {
+            // attempt to find a Scan that matches in serial, part, and date
+            Scan? Match = DbSet
+                .Where(x => x.IsIdentical(New.ToScan()))
+                .FirstOrDefault();
+            if (Match is null)
+            {
+                return true;
+            }
+            else
+            {
+                return false;
+            }
+        });
+    }
+
+    /// <summary>
+    /// Performs validation steps to ensure that New can be inserted, given the DbSet context.
+    /// </summary>
+    /// <param name="New"></param>
+    /// <param name="DbSet"></param>
+    /// <returns></returns>
+    public static async Task<ValidationFailure> Validate(ScanOutput New, IEnumerable<Scan> DbSet)
+    {
+        // validation measures from least to most expensive:
+        // -- Scan Process has NO Previous Processes
+        //    OR
+        //    Scan Process has Label Process as a Previous Process
+        //      -- if not, no Part from Previous Process is scannable
+        // -- Scan Process can scan Label Part
+        //      -- if not, this Part is not valid for Scan Process
+        // -- Label Process Required Variable Fields have non-null values
+        //      -- if not, the Label is not valid for Label Process
+        // -- Label Serial Number has been scanned by a Previous Process
+        //      -- if not, the Process was skipped
+        // confirm that the ScanOutput information is unique
+        if (!await ValidateAsUniqueScan(New, DbSet))
+        {
+            return ValidationFailure.Duplicate;
         }
-        return true;
+        // confirm that Scan Process has Label Process as a Previous Process
+        if
+        (
+            New.ScanProcess.HasPreviousProcess()
+            && !New.ScanProcess.PreviousProcesses!.Contains(New.LabelProcess.Id)
+        )
+        {
+            // Scan Process cannot accept Labels from Label Process
+            return ValidationFailure.InvalidProcess;
+        }
+        // confirm that Scan Process can scan Label Part
+        if
+        (
+            New.ScanProcess.ScanParts is null
+            || !New.ScanProcess.ScanParts.Contains(New.LabelPart.Id)
+        )
+        {
+            // Scan Process cannot accept Labels for this Part
+            return ValidationFailure.InvalidPart;
+        }
+        // confirm that all of Label Process' Required Variable Fields have values
+        if
+        (
+            New.LabelProcess.RequiredFields.JBKNumber
+            && New.LabelVariableFields.JBKNumber is null
+        )
+        {
+            return ValidationFailure.JBKNumber;
+        }
+        if
+        (
+            New.LabelProcess.RequiredFields.LotNumber
+            && New.LabelVariableFields.LotNumber is null
+        )
+        {
+            return ValidationFailure.LotNumber;
+        }
+        if
+        (
+            New.LabelProcess.RequiredFields.DieNumber
+            && New.LabelVariableFields.DieNumber is null
+        )
+        {
+            return ValidationFailure.DieNumber;
+        }
+        if
+        (
+            New.LabelProcess.RequiredFields.DeburrJBKNumber
+            && New.LabelVariableFields.DeburrJBKNumber is null
+        )
+        {
+            return ValidationFailure.DeburrJBKNumber;
+        }
+        if
+        (
+            New.LabelProcess.RequiredFields.HeatNumber
+            && New.LabelVariableFields.HeatNumber is null
+        )
+        {
+            return ValidationFailure.HeatNumber;
+        }
+        // confirm that the Label Serial # has been scanned by a Process Previous to Scan Process
+        if (!await ValidatePreviousProcessScanExists(New, DbSet))
+        {
+            return ValidationFailure.MissingPrevious;
+        }
+        // all validations passed
+        return ValidationFailure.Accepted;
     }
 }