+++ CHANGELOG.md
@@ -11,3 +11,11 @@
 - An effort to move to a CRUD API inspired design has begun.
 - Several classes have been refactored to support changes made to **LotCom Libraries**.
 - Failed network communications are now handled and do not crash the program.
+
+## `1.0.2`
+#### API Integration
+- `ScanOutput` objects utilize the API to retrieve data while parsing.
+- `NetworkService` now communicates invalid Parts and an expanded set of validation errors.
+- `ReaderService` outputs process-ready `ScanOutput` objects and reads from a new location.
+- `ScanValidationService` performs unique scan, valid process, valid part, existing previous process scan, and required field validations.
+- `Worker.cs` now operates on the last 60 days of Scans instead of the entire database (much faster).